+++ src/components/inputs/Switch/Switch.types.ts
@@ -1,7 +1,10 @@
+import type { HTMLAttributes } from "react";
 import type { InputBaseProps } from "../inputs.types";
 import type { Color } from "../../../utils/types";
 
-export interface SwitchProps extends Omit<InputBaseProps, "label"> {
+export interface SwitchProps
+  extends Omit<InputBaseProps, "label">,
+    Omit<HTMLAttributes<HTMLDivElement>, "color" | "defaultChecked" | "onChange" | "id"> {
   checked?: boolean;
   defaultChecked?: boolean;
   onCheckedChange?: (checked: boolean) => void;